pyproject.toml
[project]
name = "agentic-portfolio-analytics"
version = "0.1.0"
description = "Agentic AI platform for portfolio analytics and intelligence"
authors = [
    {name = "Tushar Vinayak", email = "tushar@example.com"},
]
dependencies = [
    "ccxt>=4.0.0",
    "pandas>=2.0.0",
    "duckdb>=0.9.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "python-dotenv>=1.0.0",
    "websockets>=11.0",
    "aiohttp>=3.8.0",
    "aiosqlite>=0.19.0",
    "greenlet>=3.0.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "fpdf2>=2.7.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0"
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = false
